--- src/basic/components/CartPage.tsx
@@ -1,134 +1,0 @@
-import { useState } from 'react';
-import { CartItem, Coupon, Product } from '../../types.ts';
-import { ProductList } from './ProductList.tsx';
-import { CartItemList } from './CartItemList.tsx';
-import { CouponApplySection } from './CouponApplySection.tsx';
-import { OrderSummary } from './OrderSummary.tsx';
-
-interface Props {
-  products: Product[];
-  coupons: Coupon[];
-}
-
-export const CartPage = ({ products, coupons }: Props) => {
-  const [cart, setCart] = useState<CartItem[]>([]);
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-
-  //1. 액션함수  
-  const addToCart = (product: Product) => {
-    const remainingStock = getRemainingStock(product);
-    if (remainingStock <= 0) return;
-
-    setCart(prevCart => {
-      const existingItem = prevCart.find(item => item.product.id === product.id);
-      if (existingItem) {
-        return prevCart.map(item =>
-          item.product.id === product.id
-            ? { ...item, quantity: Math.min(item.quantity + 1, product.stock) }
-            : item
-        );
-      }
-      return [...prevCart, { product, quantity: 1 }];
-    });
-  };
-
-  //1. 액션함수
-  const removeFromCart = (productId: string) => {
-    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
-  };
-
-  //1. 액션함수
-  const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart(prevCart =>
-      prevCart.map(item => {
-        if (item.product.id === productId) {
-          const maxQuantity = item.product.stock;
-          const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
-          return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
-        }
-        return item;
-      }).filter((item): item is CartItem => item !== null)
-    );
-  };
-
-  //계산함수
-  const calculateTotal = () => {
-    let totalBeforeDiscount = 0;
-    let totalAfterDiscount = 0;
-
-    cart.forEach(item => {
-      const { price } = item.product;
-      const { quantity } = item;
-      totalBeforeDiscount += price * quantity;
-
-      const discount = item.product.discounts.reduce((maxDiscount, d) => {
-        return quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
-      }, 0);
-
-      totalAfterDiscount += price * quantity * (1 - discount);
-    });
-
-    let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-
-    // 쿠폰 적용
-    if (selectedCoupon) {
-      if (selectedCoupon.discountType === 'amount') {
-        totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
-      } else {
-        totalAfterDiscount *= (1 - selectedCoupon.discountValue / 100);
-      }
-      totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-    }
-
-    return {
-      totalBeforeDiscount: Math.round(totalBeforeDiscount),
-      totalAfterDiscount: Math.round(totalAfterDiscount),
-      totalDiscount: Math.round(totalDiscount)
-    };
-  };
-
-  
-  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
-    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-  };
-
-  const getRemainingStock = (product: Product) => {
-    const cartItem = cart.find(item => item.product.id === product.id);
-    return product.stock - (cartItem?.quantity || 0);
-  };
-
-  const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } = calculateTotal()
-
-  const getAppliedDiscount = (item: CartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
-
-  const applyCoupon = (coupon: Coupon) => {
-    setSelectedCoupon(coupon);
-  };
-
-  return (
-    <div className="container mx-auto p-4">
-      <h1 className="text-3xl font-bold mb-6">장바구니</h1>
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <ProductList products={products} getRemainingStock={getRemainingStock} getMaxDiscount={getMaxDiscount} addToCart={addToCart} />
-        <div>
-          <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
-          <CartItemList cart={cart} getAppliedDiscount={getAppliedDiscount} updateQuantity={updateQuantity} removeFromCart={removeFromCart} />
-          <CouponApplySection coupons={coupons} applyCoupon={applyCoupon} selectedCoupon={selectedCoupon} />
-          <OrderSummary totalBeforeDiscount={totalBeforeDiscount} totalDiscount={totalDiscount} totalAfterDiscount={totalAfterDiscount} />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default CartPage;
