--- src/basic/components/index.ts
@@ -1,2 +1,0 @@
-export { default as AdminPage } from './AdminPage';
-export { default as CartPage } from './CartPage';
