+++ src/refactoring/components/coupon/CouponApplySection.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from "../../types";
+import { Coupon } from "../../../types";
 
 interface CouponApplySectionProps {
     coupons: Coupon[];
