+++ src/refactoring/components/product/ProductItemList.tsx
@@ -1,19 +1,19 @@
-import { Discount, Product } from "../../types";
+import { Discount, Product } from "../../../types";
 
 
-interface ProductListProps {
+interface ProductItemListProps {
     products: Product[];
     getRemainingStock: (product: Product) => number;
     getMaxDiscount: (discounts: Discount[]) => number;
     addToCart: (product: Product) => void;
 }
 
-export const ProductList = ({products, getRemainingStock, getMaxDiscount, addToCart}: ProductListProps) => {
+export const ProductItemList = ({products, getRemainingStock, getMaxDiscount, addToCart}: ProductItemListProps) => {
     return (
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
           <div className="space-y-2">
-            {products.map(product => {
+            {products.map((product) => {
               const remainingStock = getRemainingStock(product);
               return (
                 <div key={product.id} data-testid={`product-${product.id}`} className="bg-white p-3 rounded shadow">
